Framework/requirements.txt
flask
flask-socketio
eventlet
flask-sqlalchemy
flask-login
werkzeug
requests
pycryptodome
